+++ src/constants.js
@@ -1,0 +1,16 @@
+'use strict';
+
+module.exports = {
+  body: 'body',
+  picturesContainer: '.pictures',
+  templateElement: 'template',
+  errorElement: '.error-wrapper',
+  errorClose: '.error-close',
+  templateError: '#error-filter',
+  filters: '.filters',
+  PICTURES_LOAD_URL: 'https://o0.github.io/assets/json/pictures.json',
+  ACTIVE_FILTER_CLASSNAME: 'picture-filter-active',
+  PAGE_SIZE: 12,
+  pageNumber: 0,
+  GAP: 100
+};
