+++ src/filter/filter-type.js
@@ -1,0 +1,5 @@
+module.exports = {
+  'POPULAR': 'filter-popular',
+  'NEW': 'filter-new',
+  'DISCUSS': 'filter-discussed'
+};
