+++ src/main.js
@@ -1,5 +1,6 @@
 'use strict';
 
+require('./filter/filter');
 require('./resizer');
 require('./upload');
 require('./pictures');
