+++ src/upload.js
@@ -7,6 +7,7 @@
 
 'use strict';
 
+
 (function() {
   /** @enum {string} */
   var FileType = {
